# api/openapi.yaml
openapi: 3.0.3
info:
  title: Proxy Sampler API
  version: 1.0.0
paths:
...
  /api/config:
    get:
      operationId: runConfig
      responses:
        '200':
          description: Run-related client configuration
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunConfig'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/runs:
    post:
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRunRequest'}
      responses:
        '201':
          description: Run created and started
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    get:
      operationId: listRuns
      responses:
        '200':
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Run'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: runByID
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunDetail'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    patch:
      operationId: editRun
      description: >-
        Edits the mutable properties of a run. Renaming a run also renames the
        child sessions that still carry their generated name, leaving any
        variant renamed by hand untouched.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EditRunRequest'}
      responses:
        '200':
          description: Run updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
    delete:
      operationId: deleteRun
      responses:
        '204': {description: Run deleted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/runs/{id}/stop:
    parameters:
      - $ref: '#/components/parameters/RunID'
    post:
      operationId: stopRun
      responses:
        '200':
          description: Run stopped
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
  /api/runs/{id}/reenable:
    parameters:
      - $ref: '#/components/parameters/RunID'
    post:
      operationId: reenableRun
      responses:
        '200':
          description: Run re-enabled
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
  /api/runs/{id}/export.csv:
    parameters:
      - $ref: '#/components/parameters/RunID'
    get:
      operationId: exportRunCSV
      responses:
        '200':
          description: Deduped pool IP CSV
          headers:
            Content-Disposition: {schema: {type: string}}
          content:
            text/csv:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/DependencyUnavailable'}
...
components:
  parameters:
...
    RunID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    DependencyUnavailable:
      description: A required dependency is unavailable
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
    AxisSpec:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: {type: string, enum: [list, range, random]}
        values:
          type: array
          items: {type: string}
        from: {type: integer}
        to: {type: integer}
        count: {type: integer, minimum: 1}
        length: {type: integer, minimum: 1}
    CreateRunRequest:
      type: object
      additionalProperties: false
      required: [name, template, axes, mode, cadence_seconds]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        template: {type: string, minLength: 1, writeOnly: true}
        axes:
          type: object
          additionalProperties: {$ref: '#/components/schemas/AxisSpec'}
        mode: {type: string, enum: [sticky, pool]}
        cadence_seconds: {type: integer, minimum: 1, maximum: 2147483647}
        probes_per_sample: {type: integer, minimum: 1, maximum: 255}
        probe_target: {type: string, format: uri}
        dial_timeout_ms: {type: integer, minimum: 100, maximum: 2147483647, default: 10000}
        max_samples: {type: integer, minimum: 1, maximum: 2147483647, nullable: true}
        max_duration_seconds: {type: integer, minimum: 1, maximum: 2147483647, nullable: true}
        target_country: {type: string, pattern: '^([A-Za-z]{2})?$', nullable: true}
    RunConfig:
      type: object
      required: [max_variants_per_run]
      properties:
        max_variants_per_run: {type: integer, minimum: 1}
    EditRunRequest:
      type: object
      additionalProperties: false
      description: >-
        Mutable run properties. Every property is optional so the payload can
        grow to cover further editable fields; a request that sets none of them
        is rejected.
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
    Run:
      type: object
      required: [id, name, template_display, status, variant_count, distinct_ips, created_at]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        template_display: {type: string}
        status: {type: string, enum: [running, stopped, finished]}
        variant_count: {type: integer, minimum: 0}
        distinct_ips: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
    VariantSummary:
      type: object
      required: [session_id, name, cell_key, params, status, samples_taken, distinct_ips]
      properties:
        session_id: {type: string, format: uuid}
        name: {type: string}
        cell_key: {type: string}
        params:
          type: object
          additionalProperties: {type: string}
        target_country: {type: string, nullable: true}
        status: {type: string, enum: [running, stopped, finished]}
        samples_taken: {type: integer, minimum: 0}
        distinct_ips: {type: integer, minimum: 0}
    RunDetail:
      type: object
      required: [run, variants]
      properties:
        run: {$ref: '#/components/schemas/Run'}
        variants:
          type: array
          items: {$ref: '#/components/schemas/VariantSummary'}
...